Rename app so it doesn't have any problem with casing with fs
